requirements.txt
pandas
langgraph
langchain
langchain-community
langchain-groq
langchain-core
groq
spoonacular
langchain-chroma
IPython
transformers
sentence-transformers
kagglehub
python-dotenv
chainlit
pydantic==2.10.1
twilio==9.3.2
Flask==3.0.3